+++ server/main.go
@@ -11,14 +11,11 @@ import (
 	"flag"
 	"os"
 
-	"github.com/joho/godotenv"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
 
 func main() {
-	_ = godotenv.Load("server/.env", ".env")
-
 	serviceAccount := flag.String("fcm-service-account", "", "path to FCM service account json")
 	projectID := flag.String("fcm-project-id", "", "FCM project ID")
 	flag.Parse()